requirements.txt
# Core Data Processing
pandas>=1.5.0
numpy>=1.21.0

# Web Scraping
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
selenium>=4.5.0

# Machine Learning & NLP
sentence-transformers>=2.2.0
scikit-learn>=1.1.0
torch>=1.12.0

# String Matching
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.20.0

# PDF Processing
camelot-py>=0.10.1
opencv-python-headless>=4.6.0
ghostscript>=0.7

# Documentation Generation
weasyprint>=57.0
markdown>=3.4.0

# Utilities
python-dateutil>=2.8.0
tqdm>=4.64.0
openpyxl>=3.0.10 